requirements.txt
django
djangorestframework
django-cors-headers
uvicorn[standard]
python-dotenv
pydantic-settings
channels
daphne
